refactor: 상품 관리 기능 추가 및 상품 폼 컴포넌트 통합
- 상품 관리 기능을 위한 ProductManagement 컴포넌트 추가
- 상품 추가 및 수정 기능을 위한 ProductForm 컴포넌트 추가
- App.tsx에서 상품 관련 로직을 ProductManagement로 통합하여 코드 가독성 향상
- 상품 타입에 설명 필드 추가 및 관련 로직 수정
